US 484: get venta by id y update venta incompletos

+++ server/api/database/models/salesBranchesModel.js
@@ -20,7 +20,7 @@ SalesBranches.init(
             type: DataTypes.INTEGER
         },
         amount: {
-            type: DataTypes.BIGINT
+            type: DataTypes.FLOAT
         },
         status: {
             type: DataTypes.STRING,
